Version 2
Added buttons on the room screen to book, extend, cancel and book after current meeting.  
Added button on the room screen to go back to the flightboard.  This is made visible depending on a variable in the .env file
Added a logo to the room screen.
Added option to  hide the drop down on the flight board screen.  This is made visible depending on a variable in the .env file
Auth.js has been moved to ./config/auth/auth.js - this move creates an option to mount the folder in RAM making the auth.js file temporary.  On my own build I've encrypted auth.js and at boot if certain criteria are met, it is unencrypted and copied a ram mounted drive named auth linked to the config folder